requirements.txt
Flask==0.10.1
Flask-HTTPAuth==2.2.1
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.9.6
blist==1.3.6
cassandra-driver==2.0.2
cql==1.4.0
futures==2.1.6
itsdangerous==0.24
pycassa==1.11.0
six==1.7.3
testing.cassandra==1.1.3
thrift==0.9.1
wsgiref==0.1.2
